`refactor: 테스트용 KeywordManager SimulatedLatencyKeywordManager로 분리
